refector- move files and add option values
