+++ views/create.hbs
@@ -58,12 +58,12 @@
           <div class="row">
             <div class="col-12 mt-2 mb-5">
               <div class="d-flex justify-content-end mr-5 mb-3">
-                  <div class="d-flex mr-5">
+                  {{!-- <div class="d-flex mr-5">
                     <label for="time-range" class="d-flex justify-content-center align-items-center m-0 mr-3">Time range</label>
                     <select class="form-control" id="time-range" style="width: 15vw;">
-                      <option>Last 5 minutes</option>
-                      <option>Last 15 minutes</option>
-                      <option>Last 30 minutes</option>
+                      <option value="">Last 5 minutes</option>
+                      <option value="">Last 15 minutes</option>
+                      <option value="">Last 30 minutes</option>
                       <option>Last 1 hour</option>
                       <option>Last 3 hour</option>
                       <option>Last 6 hour</option>
@@ -78,7 +78,7 @@
                       <option>Last 2 years</option>
                       <option>Last 5 years</option>
                     </select>
-                  </div>
+                  </div> --}}
                   <button type="submit" class="btn btn-info mr-5">SAVE</button>
               </div>
               <div class="card m-auto" style="width: 90%;">
@@ -90,39 +90,40 @@
               </div>
               <div class="d-flex mt-3 ml-3">
                 <label for="source" class="d-flex justify-content-center align-items-center mx-5" style="display: block; width:3vw">Source</label>
+                
                 <select class="form-control" id="source" style="width: 15vw;">
-                  <option>collectd-cpu_value</option>
-                  <option>collectd-memory_value</option>
-                  <option>collectd-disk_value</option>
+                  {{#each source}}
+                  <option value="{{this}}">{{this}}</option>
+                  {{/each}}
                 </select>
               </div>
               <div class="d-flex">
                 <div class="d-flex mt-3 ml-3">
                   <label for="style" class="d-flex justify-content-center align-items-center mx-5" style="display: block; width:3vw">Style</label>
                   <select class="form-control" id="style" style="width: 15vw;">
-                    <option>line chart</option>
-                    <option>bar chart</option>
+                    <option value="line">line chart</option>
+                    <option value="bar">bar chart</option>
                   </select>
                 </div>
                 <div class="d-flex mt-3 ml-3">
                   <label for="interval" class="d-flex justify-content-center align-items-center mx-5" style="display: block; width:3vw">Interval</label>
-                  <input type="text" style="width: 75px;">
-                  <select class="form-control" id="unit" style="width: 5vw;" name="unit">
-                    <option>s</option>
-                    <option>m</option>
-                    <option>h</option>
-                    <option>d</option>
-                    <option>w</option>
+                  <input class="form-control" type="number" min="1" style="width: 75px;" name="interval" id="interval">
+                  <select class="form-control ml-2" style="width: 8vw;" name="interval_unit">
+                    <option value="s">second</option>
+                    <option value="m">minute</option>
+                    <option value="h">hour</option>
+                    <option value="d">day</option>
+                    <option value="w">week</option>
                   </select>
                 </div>
                 <div class="d-flex mt-3 ml-3">
                   <label for="select" class="d-flex justify-content-center align-items-center mx-5" style="display: block; width:3vw">Select</label>
                   <select class="form-control" id="select" style="width: 15vw;" name="select">
-                    <option>min()</option>
-                    <option>max()</option>
-                    <option>mean()</option>
-                    <option>sum()</option>
-                    <option>count()</option>
+                    <option value="min()">min()</option>
+                    <option value="max()">max()</option>
+                    <option value="mean()">mean()</option>
+                    <option value="sum()">sum()</option>
+                    <option value="count()">count()</option>
                   </select>
                 </div>
               </div>
